+++ src/main/java/life/inha/icemarket/IceMarketApplication.java
@@ -1,5 +1,7 @@
 package life.inha.icemarket;
 
+import life.inha.icemarket.domain.auth.User;
+import life.inha.icemarket.domain.auth.UserRepository;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
