+++ src/main/java/life/inha/icemarket/domain/auth/UserRepository.java
@@ -1,4 +1,4 @@
-package life.inha.icemarket;
+package life.inha.icemarket.domain.auth;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
