+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.env.EnvironmentPostProcessor=life.inha.icemarket.DockerSecretPostProcessor
+org.springframework.boot.env.EnvironmentPostProcessor=life.inha.icemarket.config.DockerSecretPostProcessor
