+++ src/test/java/life/inha/icemarket/domain/auth/User.java
@@ -1,0 +1,44 @@
+package life.inha.icemarket.domain.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import org.springframework.data.annotation.CreatedDate;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import java.time.LocalDateTime;
+
+@Entity(name = "users")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+    @Id
+    private int id;
+
+    @NonNull
+    private String name;
+
+    @NonNull
+    private String email;
+
+    @NonNull
+    @Column(name = "password_hashed")
+    private String passwordHashed;
+
+    private String nickname;
+
+    @CreatedDate
+    @NonNull
+    private LocalDateTime createdAt;
+
+//    @LastModifiedDate
+//    @NonNull
+//    private LocalDateTime updatedAt;
+
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt;
+}
