+++ src/test/java/life/inha/icemarket/domain/auth/UserRepository.java
@@ -1,0 +1,9 @@
+package life.inha.icemarket.domain.auth;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Integer> {
+    User findById(int id);
+}
